Introduce yg-acyclic.
This adds a new custom yield-generator script, yg-acyclic.py.  This works
very similar to the standard "basic" yield generator, except that it stops
when funds are in the "lowest" mixdepth rather than sending them back to
the initial one.

See https://github.com/JoinMarket-Org/joinmarket-clientserver/issues/404
for some discussion of this.